型の変更

+++ webapp/go/main.go
@@ -33,19 +33,19 @@ type InitializeResponse struct {
 }
 
 type Chair struct {
-	ID          int64  `db:"id" json:"id"`
+	ID          int32  `db:"id" json:"id"`
 	Name        string `db:"name" json:"name"`
 	Description string `db:"description" json:"description"`
 	Thumbnail   string `db:"thumbnail" json:"thumbnail"`
-	Price       int64  `db:"price" json:"price"`
-	Height      int64  `db:"height" json:"height"`
-	Width       int64  `db:"width" json:"width"`
-	Depth       int64  `db:"depth" json:"depth"`
+	Price       uint32  `db:"price" json:"price"`
+	Height      uint16  `db:"height" json:"height"`
+	Width       uint16  `db:"width" json:"width"`
+	Depth       uint16  `db:"depth" json:"depth"`
 	Color       string `db:"color" json:"color"`
 	Features    string `db:"features" json:"features"`
 	Kind        string `db:"kind" json:"kind"`
-	Popularity  int64  `db:"popularity" json:"-"`
-	Stock       int64  `db:"stock" json:"-"`
+	Popularity  int32  `db:"popularity" json:"-"`
+	Stock       int32  `db:"stock" json:"-"`
 }
 
 type ChairSearchResponse struct {
@@ -59,18 +59,18 @@ type ChairListResponse struct {
 
 //Estate 物件
 type Estate struct {
-	ID          int64   `db:"id" json:"id"`
+	ID          int32   `db:"id" json:"id"`
 	Thumbnail   string  `db:"thumbnail" json:"thumbnail"`
 	Name        string  `db:"name" json:"name"`
 	Description string  `db:"description" json:"description"`
 	Latitude    float64 `db:"latitude" json:"latitude"`
 	Longitude   float64 `db:"longitude" json:"longitude"`
 	Address     string  `db:"address" json:"address"`
-	Rent        int64   `db:"rent" json:"rent"`
-	DoorHeight  int64   `db:"door_height" json:"doorHeight"`
-	DoorWidth   int64   `db:"door_width" json:"doorWidth"`
+	Rent        int32   `db:"rent" json:"rent"`
+	DoorHeight  uint16   `db:"door_height" json:"doorHeight"`
+	DoorWidth   uint16   `db:"door_width" json:"doorWidth"`
 	Features    string  `db:"features" json:"features"`
-	Popularity  int64   `db:"popularity" json:"-"`
+	Popularity  int32   `db:"popularity" json:"-"`
 }
 
 //EstateSearchResponse estate/searchへのレスポンスの形式
